refactor: domain, controller, view 패키지 생성
