+++ src/test/java/com/mission2/racingcar/domain/CarTests.java
@@ -1,4 +1,4 @@
-package com.mission2.racingcar;
+package com.mission2.racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
